+++ admin/admin_login.php
@@ -5,7 +5,7 @@
 </script>
 
 <?php
-    require_once 'db_connection.php';
+    require_once '../partials/db_connection.php';
 
     if(isset($_SESSION['ad_id']))
     {
@@ -55,25 +55,25 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
-    <link rel="stylesheet" href="fontawesome-6/css/all.css">
-    <link rel="stylesheet" href="css/admin_login.css">
-    <link rel="stylesheet" href="css/nav.css">
-    <link rel="stylesheet" href="css/footer.css">
+    <link rel="stylesheet" href="../fontawesome-6/css/all.css">
+    <link rel="stylesheet" href="../css/admin_login.css">
+    <link rel="stylesheet" href="../css/nav.css">
+    <link rel="stylesheet" href="../css/footer.css">
     
 
     <title>admin login</title>
-    <link rel="icon" type="image/x-icon" href="favicon.ico">
+    <link rel="icon" type="image/x-icon" href="../favicon.ico">
 </head>
 <body>
 
-    <?php require_once 'partials/nav.php' ?>
+    <?php require_once '../partials/nav.php' ?>
 
     <div class="container">
         <div class="login_box">
             <form action="" method="POST">
                 <div class="icon">
                     <!-- <i class="fa-solid fa-user-large"></i> -->
-                    <img src="images/user_admin4.svg" alt="">
+                    <img src="../images/user_admin4.svg" alt="">
                     <p>admin login</p>
                 </div>
                 <div class="fields">
@@ -94,7 +94,7 @@
         </div>
     </div>
 
-    <?php require_once 'partials/footer.php' ?>
+    <?php require_once '../partials/footer.php' ?>
     
 </body>
 </html>
