+++ partials/nav_index.php
@@ -1,13 +1,13 @@
 <div class="navigration">
         <div class="logo">
             <a href="index.php">
-                <img src="../images/logo.svg" alt="logo">
+                <img src="images/logo.svg" alt="logo">
             </a>
         </div>
         <div class="nav">
             <nav>
                 <ul>
-                    <li><a href="../index.php">home</a></li>
+                    <li><a href="index.php">home</a></li>
                     <li><a href="#">catagory</a>
                         <ul class="nested">
                             <li><a href="#">thesis</a></li>
@@ -16,8 +16,8 @@
                     </li>
                     <li><a href="#">login</a>
                         <ul class="nested">
-                            <li><a href="../supervisor/supervisor_login.php">supervisor</a></li>
-                            <li><a href="../admin/admin_login.php">admin</a></li>
+                            <li><a href="supervisor/supervisor_login.php">supervisor</a></li>
+                            <li><a href="admin/admin_login.php">admin</a></li>
                         </ul>
                     </li>
                     <li><a href="#">about</a></li>
